third_party/blink/web_tests/clipboard/async-clipboard/async-unsanitized-standard-html-formats-write-read.html

<!doctype html>
<meta charset="utf-8">
<title>Async Clipboard unsanitized HTML write -> Async Clipboard unsanitized HTML read test</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="resources/user-activation.js"></script>
<script>
'use strict';
// This function removes extra spaces between tags in html. For example, the
// following html: "<p> Hello </p>   <body> World </body>" would turn into this
//           html: "<p> Hello </p> <body> World </body>"
// We remove the extra spaces because in html they are considered equivalent,
// but when we are comparing for equality the spaces make a difference.
function reformatHtml(html) {
  return html.replace(/\>\s*\</g, '> <').replace(/\<!--[a-zA-Z]*?--\>/g, '');
}

async function readWriteTest(textInput) {
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');

  // Create and write unsanitized version of standard HTML format.
  const format1 = 'text/html';
  const blobInput1 = new Blob([textInput.input], {type: format1});
  const clipboardItemInput = new ClipboardItem({[format1]: blobInput1});
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItemInput]);

  // Read unsanitized version of HTML format.
  await waitForUserActivation();
  const clipboardItems =
    await navigator.clipboard.read({unsanitized: ["text/html"]});
  assert_equals(clipboardItems.length, 1);
  const clipboardItem = clipboardItems[0];
  assert_true(clipboardItem instanceof ClipboardItem);

  const blobOutput1 = await clipboardItem.getType(format1);
  assert_equals(blobOutput1.type, format1);

  const data1 = await (new Response(blobOutput1)).text();
  const outputHtml = reformatHtml(data1);
  if ("expected" in textInput) {
    assert_equals(outputHtml, reformatHtml(textInput.expected));
  } else {
    assert_equals(outputHtml, reformatHtml(textInput.input));
  }
}

// The string must be concatenated in this way because the html parser
// will recognize a script tag even in quotes as a real script tag. By
// splitting it up in this way we avoid that error.
const html_script = '<script>const a = 5;</scr'
                          + 'ipt>';
const startFragment = '<!--StartFragment -->';
const endFragment = '<!--EndFragment -->';
const testCases = [
  { input:    '<head><style>color:blue</style></head>' +
              '<body>' + html_script + '<p>Hello World</p></body>',
    expected: '<html><head><style>color:blue</style></head>' +
              '<body>' + html_script + '<p>Hello World</p></body></html>' },
  { input:    '<head><style>color:blue</style>' +
              html_script + '</head>' +
              '<body>' + startFragment + '<p>Hello</p>' +
              endFragment + '</body>',
    expected: '<html><head><style>color:blue</style>' +
              html_script + '</head>' +
              '<body>' + startFragment + '<p>Hello</p>' +
              endFragment + '</body></html>' },
];

promise_test(async t => {
  for (const testCase of testCases) {
    await readWriteTest(testCase);
  }
}, 'Verify read and write of unsanitized content to the clipboard given text/html format as input');
</script>
